l4p/tcp: multithreaded tle_tcp_process can lock up
We believe this is the correct set of step to follow this issue

1 - BE: RX a reset packet and start the rx_bulk call, we get past
rx_acquire_stream call but not all the way through the code path to
resetting the stream variables and critically not to the point of
setting type = TLE_VNUM. The rx_acquire_stream locks the stream, which
should prevent critical modifications.
2 - FE: Call tle_tcp_stream_abort, which calls stream_close_prolog with
operation ABORT_CLOSE
2a - FE: tle_tcp_stream_abort sets the stream operation for
stream_close_prolog to TLE_TCP_OP_CLOSE|TLE_TCP_OP_ABORT and then in
stream_close_prolog calls tcp_stream_down, pause processing in that
thread here.

This pause isn't just incidental, it is enforced by locking when this
race hits the right sequencing. The BE rx_acquire_stream forces a pause
in FE tcp_stream_down because it tries to lock as well, and is blocked
by the BE holding the lock.

3 - BE: continues processing the received RST through rx_stream which
will call into rx_rst; we're at a place the RST would pass check_seqn so
stream_term would get called
3a - BE: rx_rst calls stream_term which sets s->tcb.state =
TLE_TCP_ST_CLOSED; and because tle_tcp_stream_abort set
TLE_TCP_OP_CLOSE, the stream is enqueued to DR
3b - BE: back in rx_bulk the BE releases the lock by calling
tcp_stream_release allowing the FE to make progress
4 - FE: Picks back up in stream_close_prolog; now that the BE has set
stream state to closed, state <= TLE_TCP_ST_SYN_SENT is true in
stream_close_prolog so tcp_stream_reset(ctx, s); is called then return
0; causes tle_tcp_stream_abort to skip the tcp_stream_up
5 - BE: Back in the BE tle_tcp_process runs the death row list, which
our stream is now on and down, and calls tcp_stream_down then
tcp_stream_reset

----

The moved line (the stream down) puts the update to s->tcb.uop inside a
critical section protecting it from simultaneous update. When the FE
gets to step 2 it pauses and waits on the lock. The FE doesn't get to
2a. So the stream operation is NOT set to
TLE_TCP_OP_CLOSE|TLE_TCP_OP_ABORT yet which means in the BE in 3a the
check does NOT put the stream in the DR and instead raises an event.

Once the BE finishes up and unlocks the FE picks back up. THEN it
updates s->tcb.uop in stream_close_prolog and keeps processing and picks
up at 4. But step 5 is never run, because the sequencing did not trigger
death row enqueue.

Signed-off-by: Ben Magistro <koncept1@gmail.com>
Change-Id: Idffe9fca7ebf68f1fef38ecaa163e7c2f55e43f5