l4p/tcp: prevent send window underflow
Due to quantization of the TCP window with the scaling factor, there are cases
where more data may be in flight than the received advertised TCP window. This
is not a bug and is normal behavior. It does mean that math on in flight data
and the advertised receive window can sometimes produce negative or underflow
values that need to saturate at 0.

Private Issue Tracking: PS-3155