l4p/tcp: minimize impact of quantized rx buffer
Because of TLDK's quantized receive buffer, there are cases where a received
segment reduces the advertised TCP window by more than the data in the packet.
When this happens, there is a chance the sender put data in flight which no
longer can be received because TLDK is out of mbufs to store the data. There
are two major cases seen so far:

Case 1 - large transfer, end of window partial packets:
(This patch addresses this case.)
One common edge case for this is at the end of a full window transmission,
often the window being advertised becomes small enough for partial packets to
be sent. This over-shirnks the window size and data in flight towards the end
of a full window gets lost.

This patch reserves a bit of space to not advertise in sent packets, but to
keep in the TCP window computations internally. This allows frames to arrive
and be accepted if they are in the window space between the end of the
advertised window and the end of the internal window. But, the sender will not
explicitly try to fill that space intentionally. When a few small frames are
sent towards the end of a window, there will be a few mbufs reserved to accept
them and not loose data.

Case 2 - many small frames:
(This case is out of scope of this patch, but is worth mentioning.)
In cases of many smaller than estimated segments, there is not much that can
be done other than reduce the estimated size, which reduces the window overall
and throughput. However, throughput is already reduced for these cases where
frames are not fully utilizing the available mbuf space.

This is more common in cases of interactive protocols less likely to fill
the entire TCP window, if a sufficient number of mbufs are available. There
may be protocols that have an "internal" MTU that limits total TCP segment
size but still attempt to transfer large amounts of data. This is unfortunate
and not easily handled without a non-quantized receive buffer.

Private Issue Tracking: PS-3160