l4p/tcp: avoid overwriting mbufs as they are transmitted
When retransmitting an mbuf, TLDK updates an existing mbuf and increments the
refcnt. For the first retransmission, this is usually fine as the original
likely has already been sent. For a retransmission of the retransmission it
is possible the mbuf is still in the transmit queue and has not been sent
yet when a second retransmit triggers. This allows updates to the mbuf to be
made before or while the mbuf is sent causing corruption.

Copying mbufs for retransmissions avoids this problem.

Private Issue Tracking: PS-3159