l4p/tcp: send window updates when window opens from 0
When the TCP window is advertised as 0 and the front end reads data out of the
RX queue, opening the window up to a non-0 advertised value, the sender is
unaware and does not know to start transmitting again. A TCP window update can
and should be sent when transitioning from advertised 0 to advertised non-0 RX
window.

Note: this is about advertised 0, not actual 0. An actual window of 2048 but
with a scaling factor of 8192 may get advertised as 0 until 8193 or more bytes
are available in the queue. The transition to window update on is NOT true 0
to non-0 but advertised 0 to non-advertised 0.

Private Issue Tracking: PS-3161